Basic turn order set up.
Implemented a basic version of the turn order. When it is a players turn, they can press 'roll dice' to call the rollAndMove function, which will then create an alert showing what they rolled and where they are now. After that, they can press the 'end turn' button to end their turn, which will send an alert to the next player that it is their turn. If it is not a players turn yet, the 'roll dice' and 'end turn' buttons will alert the player it is not their turn yet.